Move composition root out of core into bootstrap package
