+++ bootstrap/meta.py
@@ -1,0 +1,9 @@
+import os
+
+
+def name():
+    return "tomsg_bot"
+
+
+def version():
+    return os.getenv("VERSION", "undefined")
